Исправит замечания.
