+++ src/utils.js
@@ -1,4 +1,4 @@
-const capitalizeLetter = (someString) => (someString[0].toUpperCase() + someString.slice(1));
+const capitalizeLetter = (someString) => someString[0].toUpperCase() + someString.slice(1);
 
 export {capitalizeLetter};
