+++ src/view/loading-view/loading-view.js
